+++ DMath.h
@@ -1,5 +1,5 @@
 #pragma once
 class DMath {
 public:
-	static bool IsEqual(double Lhs, double Rhs);
+	static bool IsEqual(float Lhs, float Rhs);
 };
