+++ MathHelper.cpp
@@ -11,25 +11,25 @@ Vector2 MathHelper::GetGradientVec(const Vector2& A, const Vector2& B) {
 	return {A.GetY() - B.GetY(), B.GetX() - A.GetX()};
 }
 
-double MathHelper::LerpD(const double& A, const double& B, const float Ratio) {
-	return (B - A) * static_cast<double>(Ratio) + A;
+float MathHelper::LerpD(const float& A, const float& B, const float Ratio) {
+	return (B - A) * static_cast<float>(Ratio) + A;
 }
 
-float MathHelper::LerpF(const double& A, const double& B, const float Ratio) {
-	return static_cast<float>((B - A) * static_cast<double>(Ratio) + A);
+float MathHelper::LerpF(const float& A, const float& B, const float Ratio) {
+	return static_cast<float>((B - A) * static_cast<float>(Ratio) + A);
 }
 
-uint32_t MathHelper::Floor(const double A) { return static_cast<uint32_t>(A); }
+uint32_t MathHelper::Floor(const float A) { return static_cast<uint32_t>(A); }
 
-double MathHelper::SolveLine(const double& CurX, const Vector2& P1, const Vector2& P2) {
-	const double a = P1.GetY() - P2.GetY();
-	const double b = P2.GetX() - P1.GetX();
-	const double c = (P1.GetX() - P2.GetX()) * P1.GetY() + (P2.GetY() - P1.GetY()) * P1.GetX();
+float MathHelper::SolveLine(const float& CurX, const Vector2& P1, const Vector2& P2) {
+	const float a = P1.GetY() - P2.GetY();
+	const float b = P2.GetX() - P1.GetX();
+	const float c = (P1.GetX() - P2.GetX()) * P1.GetY() + (P2.GetY() - P1.GetY()) * P1.GetX();
 
 	return (a * CurX + c) / b;
 }
 
-double MathHelper::SolveLine2(const double& X, const Vector2& P1, const Vector2& P2) {
+float MathHelper::SolveLine2(const float& X, const Vector2& P1, const Vector2& P2) {
 	Vector2 max;
 	Vector2 min;
 	GetMinAndMaxX(min, max, P1, P2);
@@ -37,7 +37,7 @@ double MathHelper::SolveLine2(const double& X, const Vector2& P1, const Vector2&
 	return (P2.GetY() - P1.GetY()) / (P2.GetX() - P1.GetX()) * (X - P1.GetX()) + P1.GetY();
 }
 
-uint32_t MathHelper::Get2DIndex(const double& X, const double& Y, const uint32_t BufferWidth) {
+uint32_t MathHelper::Get2DIndex(const float& X, const float& Y, const uint32_t BufferWidth) {
 	return static_cast<uint32_t>(Y * BufferWidth + X);
 }
 
@@ -88,11 +88,11 @@ void MathHelper::GetMinAndMaxY(Vector2& Min, Vector2& Max, const Vector2& Start,
 //	return {clip_space_x, clip_space_y};
 //}
 
-double MathHelper::Sqrt(const double square) {
-	/*double root = square / 3, diff = 1;
+float MathHelper::Sqrt(const float square) {
+	/*float root = square / 3, diff = 1;
 	if (square <= 0) return 0;
 	do {
-		const double last = root;
+		const float last = root;
 		root = (root + square / root) / 2;
 		diff = root - last;
 	}
@@ -101,7 +101,7 @@ double MathHelper::Sqrt(const double square) {
 	return std::sqrt(square);
 }
 
-double MathHelper::Cot(const double& A) { return std::cos(A) / std::sin(A); }
+float MathHelper::Cot(const float& A) { return std::cos(A) / std::sin(A); }
 
-double MathHelper::Deg2Rad(const double& Deg) { return Deg * PI / 180; }
+float MathHelper::Deg2Rad(const float& Deg) { return Deg * PI / 180; }
