+++ Vector2.h
@@ -5,7 +5,7 @@ class Vector2 {
 public:
 	/**=== Constructors ===*/
 	Vector2();
-	Vector2(double X, double Y);
+	Vector2(float X, float Y);
 	~Vector2();
 
 	/**=== Copy and Move ===*/
@@ -20,24 +20,24 @@ public:
 	Vector2 operator-(const Vector2& Other) const;
 	Vector2& operator-=(const Vector2& Other);
 
-	Vector2 operator*(double Other) const;
-	Vector2& operator*=(double Other);
+	Vector2 operator*(float Other) const;
+	Vector2& operator*=(float Other);
 
 	/**=== Getters and Setters ===*/
-	double GetX() const;
-	double GetY() const;
+	float GetX() const;
+	float GetY() const;
 
-	void SetX(double NewX);
-	void SetY(double NewY);
+	void SetX(float NewX);
+	void SetY(float NewY);
 
 	/** Get the z component for the vector 2, this is the homogeneous coordinate and would be used for depth. */
-	double GetZ() const;
-	void SetZ(double Z);
+	float GetZ() const;
+	void SetZ(float Z);
 
-	double SquaredMagnitude() const;
+	float SquaredMagnitude() const;
 
 	Vector2& Translate(const Vector2& Other);
-	Vector2& Scale(double Multiplier);
+	Vector2& Scale(float Multiplier);
 	Vector2& Scale(const Vector2& Other);
 
 	friend bool operator==(const Vector2& Lhs, const Vector2& Rhs);
@@ -47,5 +47,5 @@ public:
 
 private:
 	/** Array of */
-	double Position[3]; // XYZ
+	float Position[3]; // XYZ
 };
